fix: 修复LED闪烁干扰和input_state数据类型错误
Bug1: UI系统LED分配与继电器状态灯冲突

- UI LED统一改为LED1(P0_0)，不干扰继电器LED2-4(P0_1~P0_3)

- UI_DeviceStateUpdated入网后停止LED1闪烁，未配网时闪烁提示

- ZDO_STATE_CHANGE后恢复LED1的继电器状态显示

- RelayInit中初始化LED为继电器OFF状态(亮)

Bug2: genAnalogInput.presentValue数据类型不匹配

- ZCL标准类型为single_float(0x39)，原误用UINT8(0x20)

- Z2M写入时返回INVALID_DATA_TYPE

- InputState改为float类型，添加WRITE权限

同时添加.gitignore忽略zigbee-herdsman-converters目录

+++ Projects/zstack/HomeAutomation/SampleSwitch/Source/zcl_samplesw.h
@@ -109,7 +109,7 @@ extern CONST zclAttrRec_t zclSampleSw_Attrs[];
 
 extern uint8 zclSampleSw_RelayState[SAMPLESW_NUM_RELAYS];
 
-extern uint8 zclSampleSw_InputState[SAMPLESW_NUM_INPUTS];
+extern float zclSampleSw_InputState[SAMPLESW_NUM_INPUTS];
 
 extern uint8  zclSampleSw_OnOff;
